Optimize BitOps.ConditionalSetBit

+++ SharpBlaze/LineArrayTiled.Inlines.cs
@@ -536,7 +536,7 @@ public unsafe partial struct LineArrayTiled<T>
         TileIndex columnIndex, F24Dot8 x0, F24Dot8 y0,
         F24Dot8 x1, F24Dot8 y1)
     {
-        if (BitOps.ConditionalSetBit(mBitVectors, (int) columnIndex))
+        if (BitOps.ConditionalSetBit(mBitVectors, columnIndex))
         {
             // First time line is inserted into this column.
             LineArrayTiledBlock* b = memory.FrameNewTiledBlock(null);
